[Github #672] improvements (#679)
* [gh-#672] improve error response if the connection isn't found

* [gh-#672] show the provider config key

* [gh-#672] update docs for Base-Url-Override

* [gh-#672] set process directly to assume cloud

•

* [gh-#672] better cli example that uses github and cleaner output of results

* [gh-#672] add retries to the activity log

* [gh-#672] sync commands go to the activity log

* [gh-#672] dev: restart server and worker on .env change

* [gh-#672] sync command to activity log

* [gh-#672] schedule update goes to error

* [gh-#672] prettier batch send log if dry run

* [gh-#672] sync command to activity log

* [gh-#672] docs update to await batchSend

* [gh-#672] send models instead of a single model each time

* [gh-#672] error handling

* [gh-#672] updat ejs file

* [gh-#672] working example for cli integration template

* [gh-#672] verify the batchSend sync model matches what is in the sync job

* [gh-#672] better end report based on the added results

* [gh-#672] better unknown provider config error report

* [gh-#672] use payload in the error response

* [gh-#672] don't run as background

* [gh-#672] 0.21.4

* [gh-#672] webhook update to return the job updated at timestamp

* [gh-#672] create models file if it doesn't exist on compile

• add version command

* [gh-#672] add upgrade logic and improve docker down via control c

* [gh-#672] don't deploy if no integration files and add checks for necessary files

* [gh-#672] warn await calls for integration files

* [gh-#672] useEffect for default scope

* [gh-#672] 0.21.5

* [gh-#672] check types

* [gh-#672] 0.21.9

* [gh-#672] allow option to not overwrite the interfaces when compiling in dev mode

* [gh-#672] safer log showing

* [gh-#672] config and model loading fixes

* [gh-#672] dynamic port for cli produced application, config deploy start logic, and verify file function

* [gh-#672] start reconcile sync logic

* [gh-#672] expose connectionId and providerConfigKey

* [gh-#672] allow NANGO_INTEGRATIONS_LOCATION to be set in the application

* [gh-#672] set isSync by default

* [gh-#672] update env variable for nango integrations app

* [gh-#672] fix if futureActionTimes is undefined

* [gh-#672] log if a webhook isn't sent

* [gh-#672] parse sync results based on the json structure

* [gh-#672] store sync models by a model key

* [gh-#672] reconcile updates

* [gh-#672] update logic for last sync date

* [gh-#672] deleting a connection deletes the schedule

* [gh-#672] update env variables

* [gh-#672] don't dynamically update syncs for now

* [gh-#672] safer error message parsing

* [gh-#672] bump version

+++ packages/cli/lib/sync.ts
@@ -1,6 +1,7 @@
 import fs from 'fs';
 import path, { dirname } from 'path';
 import { fileURLToPath } from 'url';
+import axios from 'axios';
 import yaml from 'js-yaml';
 import chalk from 'chalk';
 import chokidar from 'chokidar';
@@ -12,11 +13,23 @@ import { spawn } from 'child_process';
 import parser from '@babel/parser';
 import traverse, { NodePath } from '@babel/traverse';
 import type { ChildProcess } from 'node:child_process';
+import promptly from 'promptly';
 import type * as t from '@babel/types';
 
 import type { NangoConfig, Connection as NangoConnection, NangoIntegration, NangoIntegrationData } from '@nangohq/shared';
-import { cloudHost, stagingHost, SyncType, syncRunService, nangoConfigFile } from '@nangohq/shared';
-import { hostport, getConnection, NANGO_INTEGRATIONS_LOCATION, buildInterfaces } from './utils.js';
+import { loadSimplifiedConfig, cloudHost, stagingHost, SyncType, syncRunService, nangoConfigFile, checkForIntegrationFile } from '@nangohq/shared';
+import {
+    port,
+    hostport,
+    httpsAgent,
+    verifyNecessaryFiles,
+    getConnection,
+    NANGO_INTEGRATIONS_LOCATION,
+    buildInterfaces,
+    enrichHeaders,
+    checkEnvVars
+} from './utils.js';
+import type { DeployOptions } from './types.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -34,22 +47,175 @@ interface RunArgs {
 const exampleSyncName = 'github-issue-example';
 
 const createModelFile = (notify = false) => {
-    if (!fs.existsSync(`${NANGO_INTEGRATIONS_LOCATION}/models.ts`)) {
-        const cwd = process.cwd();
-        const configContents = fs.readFileSync(path.resolve(cwd, `${NANGO_INTEGRATIONS_LOCATION}/${nangoConfigFile}`), 'utf8');
-        const configData: NangoConfig = yaml.load(configContents) as unknown as NangoConfig;
-        const { models } = configData;
-        const interfaceDefinitions = buildInterfaces(models);
-        fs.writeFileSync(`${NANGO_INTEGRATIONS_LOCATION}/models.ts`, interfaceDefinitions.join('\n'));
+    const cwd = process.cwd();
+    const configContents = fs.readFileSync(path.resolve(cwd, `${NANGO_INTEGRATIONS_LOCATION}/${nangoConfigFile}`), 'utf8');
+    const configData: NangoConfig = yaml.load(configContents) as unknown as NangoConfig;
+    const { models } = configData;
+    const interfaceDefinitions = buildInterfaces(models);
+    fs.writeFileSync(`${NANGO_INTEGRATIONS_LOCATION}/models.ts`, interfaceDefinitions.join('\n'));
 
-        if (notify) {
-            console.log(
-                chalk.green(
-                    `${NANGO_INTEGRATIONS_LOCATION}/${nangoConfigFile} was updated. The interface file (${NANGO_INTEGRATIONS_LOCATION}/models.ts) was updated to reflect the updated config`
-                )
-            );
+    if (notify) {
+        const rawNangoIntegrationLocation = NANGO_INTEGRATIONS_LOCATION.replace('./', '');
+        console.log(
+            chalk.green(
+                `${rawNangoIntegrationLocation}/${nangoConfigFile} was updated. The interface file (${rawNangoIntegrationLocation}/models.ts) was updated to reflect the updated config`
+            )
+        );
+    }
+};
+
+const getConfig = async () => {
+    const cwd = process.cwd();
+    const config = await loadSimplifiedConfig(path.resolve(cwd, NANGO_INTEGRATIONS_LOCATION));
+
+    if (!config) {
+        throw new Error(`Error loading the ${nangoConfigFile} file`);
+    }
+
+    return config;
+};
+
+const deployLocalSyncs = async () => {
+    console.log(chalk.green(`Syncing the updated config so syncs will be created or deleted according to the changed config`));
+
+    const url = hostport + `/sync/reconcile`;
+    const config = await getConfig();
+
+    const postData = [];
+
+    const cwd = process.cwd();
+
+    for (const integration of config) {
+        const { providerConfigKey } = integration;
+        const { syncs } = integration;
+
+        for (const sync of syncs) {
+            const { name: syncName, runs, returns } = sync;
+
+            const { result: integrationFileResult } = checkForIntegrationFile(syncName, path.resolve(cwd, `${NANGO_INTEGRATIONS_LOCATION}`));
+
+            if (!integrationFileResult) {
+                continue;
+            }
+
+            const body = {
+                syncName,
+                providerConfigKey,
+                runs,
+                returns
+            };
+
+            postData.push(body);
         }
     }
+
+    await axios
+        .post(url, postData, { headers: enrichHeaders(), httpsAgent: httpsAgent() })
+        .then((response: any) => {
+            console.log(chalk.green(`Syncs were updated with the following result: ${JSON.stringify(response.data, null, 2)}`));
+        })
+        .catch((err) => {
+            const errorMessage = JSON.stringify(err?.response?.data, null, 2);
+            console.log(chalk.red(`Error updating the syncs with the following error: ${errorMessage}}`));
+            process.exit(1);
+        });
+};
+
+export const deploy = async (options: DeployOptions) => {
+    const { staging, version, sync: optionalSyncName, secretKey, host } = options;
+    await verifyNecessaryFiles();
+
+    if (host) {
+        process.env['NANGO_HOSTPORT'] = host;
+    }
+
+    if (secretKey) {
+        process.env['NANGO_SECRET_KEY'] = secretKey;
+    }
+
+    if (!process.env['NANGO_HOSTPORT']) {
+        if (staging) {
+            process.env['NANGO_HOSTPORT'] = stagingHost;
+        } else {
+            process.env['NANGO_HOSTPORT'] = cloudHost;
+        }
+    }
+
+    if (process.env['NANGO_HOSTPORT'] !== `http://localhost:${port}` && !process.env['NANGO_SECRET_KEY']) {
+        console.log(chalk.red(`NANGO_SECRET_KEY environment variable is not set. Please set it now`));
+        try {
+            const secretKey = await promptly.prompt('Secret Key: ');
+            if (secretKey) {
+                process.env['NANGO_SECRET_KEY'] = secretKey;
+            } else {
+                return;
+            }
+        } catch (error) {
+            console.log('Error occurred while trying to prompt for secret key:', error);
+            process.exit(1);
+        }
+    }
+
+    checkEnvVars(process.env['NANGO_HOSTPORT']);
+    tsc();
+
+    const config = await getConfig();
+    const cwd = process.cwd();
+
+    const postData = [];
+
+    for (const integration of config) {
+        const { providerConfigKey } = integration;
+        let { syncs } = integration;
+
+        if (optionalSyncName) {
+            syncs = syncs.filter((sync) => sync.name === optionalSyncName);
+        }
+
+        for (const sync of syncs) {
+            const { name: syncName, runs, returns: models, models: model_schema } = sync;
+
+            const { path: integrationFilePath, result: integrationFileResult } = checkForIntegrationFile(
+                syncName,
+                path.resolve(cwd, `${NANGO_INTEGRATIONS_LOCATION}`)
+            );
+
+            if (!integrationFileResult) {
+                console.log(chalk.red(`No integration file found for ${syncName} at ${integrationFilePath}. Skipping...`));
+                continue;
+            }
+
+            const body = {
+                syncName,
+                providerConfigKey,
+                models,
+                version,
+                runs,
+                fileBody: fs.readFileSync(integrationFilePath, 'utf8'),
+                model_schema: JSON.stringify(model_schema)
+            };
+
+            postData.push(body);
+        }
+    }
+
+    const url = process.env['NANGO_HOSTPORT'] + `/sync/deploy`;
+
+    if (postData.length === 0) {
+        console.log(chalk.red(`No syncs found to deploy. Please make sure your integration files compiled successfully and exist in your dist directory`));
+        return;
+    }
+
+    await axios
+        .post(url, postData, { headers: enrichHeaders(), httpsAgent: httpsAgent() })
+        .then((_) => {
+            console.log(chalk.green(`Successfully deployed the syncs!`));
+        })
+        .catch((err) => {
+            const errorMessage = JSON.stringify(err.response.data, null, 2);
+            console.log(chalk.red(`Error deploying the syncs with the following error: ${errorMessage}}`));
+            process.exit(1);
+        });
 };
 
 export const version = () => {
@@ -119,6 +285,7 @@ export const init = () => {
         fs.writeFileSync(
             '.env',
             `#NANGO_HOSTPORT=https://api-staging.nango.dev
+#NANGO_AUTO_UPGRADE=true
 #NANGO_SECRET_KEY=xxxx-xxx-xxxx
 #NANGO_INTEGRATIONS_LOCATION=use-this-to-override-where-the-nango-integrations-directory-goes
 #NANGO_PORT=use-this-to-override-the-default-3003
@@ -181,13 +348,13 @@ export const generate = async () => {
 };
 
 export const run = async (args: string[], options: RunArgs) => {
-    let syncName, providerConfigKey, connectionId, suppliedLastSyncDate, useServerLastSyncDate;
+    let syncName, providerConfigKey, connectionId, suppliedLastSyncDate;
     if (args.length > 0) {
-        [syncName, providerConfigKey, connectionId, suppliedLastSyncDate, useServerLastSyncDate] = args;
+        [syncName, providerConfigKey, connectionId, suppliedLastSyncDate] = args;
     }
 
     if (Object.keys(options).length > 0) {
-        ({ sync: syncName, provider: providerConfigKey, connection: connectionId, lastSyncDate: suppliedLastSyncDate, useServerLastSyncDate } = options);
+        ({ sync: syncName, provider: providerConfigKey, connection: connectionId, lastSyncDate: suppliedLastSyncDate } = options);
     }
 
     if (!syncName) {
@@ -223,10 +390,6 @@ export const run = async (args: string[], options: RunArgs) => {
 
     if (suppliedLastSyncDate) {
         lastSyncDate = new Date(suppliedLastSyncDate as string);
-    }
-
-    if (!useServerLastSyncDate) {
-        lastSyncDate = null;
     }
 
     const cwd = process.cwd();
@@ -349,6 +512,9 @@ export const tscWatch = () => {
     });
 
     watcher.on('unlink', (filePath: string) => {
+        if (filePath === `${rawNangoIntegrationLocation}/${nangoConfigFile}`) {
+            return;
+        }
         const jsFilePath = path.join(path.dirname(filePath), path.basename(filePath, '.ts') + '.js');
         const distJSFilePath = jsFilePath.replace(rawNangoIntegrationLocation, `${rawNangoIntegrationLocation}/dist`);
         fs.unlinkSync(distJSFilePath);
@@ -393,12 +559,29 @@ export const configWatch = () => {
     const watchPath = `${NANGO_INTEGRATIONS_LOCATION}/${nangoConfigFile}`;
     const watcher = chokidar.watch(watchPath, { ignoreInitial: true });
 
-    watcher.on('add', () => {
-        createModelFile(true);
-    });
-
     watcher.on('change', () => {
         createModelFile(true);
+    });
+};
+
+/**
+ * Config Deploy
+ * @desc if the config is changed and a new sync is added or deleted, make sure
+ * the worker runs the new sync or removes that new sync
+ * so any changes are seamless
+ */
+let isDeploying = false;
+
+export const configDeploy = async () => {
+    const watchPath = `${NANGO_INTEGRATIONS_LOCATION}/${nangoConfigFile}`;
+    const watcher = chokidar.watch(watchPath, { ignoreInitial: true });
+
+    watcher.on('all', async (event: string) => {
+        if (!isDeploying && (event === 'add' || event === 'change')) {
+            isDeploying = true;
+            await deployLocalSyncs();
+            isDeploying = false;
+        }
     });
 };
